requirements.txt
langchain-core
langchain-openai
langchain-google-genai
google-adk
google-genai
python-dotenv
